Fromdigits would call GMP with bases or arrays it couldn't support.  fromdigits would allow holes in array references.  fromdigits lengths and bases could overflow.  fromdigits would segfault with a sparse array reference.  Similar but unrelated: vecsample would segfault if given a sparse array reference.

+++ util.h
@@ -134,8 +134,8 @@ extern char* pidigits(uint32_t digits);
 
 
 extern bool from_digit_string(UV* n, const char* s, int base);
-extern bool from_digit_to_UV(UV* rn, const UV* r, int len, int base);
-extern bool from_digit_to_str(char** rstr, const UV* r, int len, int base);
+extern bool from_digit_to_UV(UV* rn, const UV* r, size_t len, int base);
+extern bool from_digit_to_str(char** rstr, const UV* r, size_t len, int base);
 /* These return length */
 extern int  to_digit_array(int* bits, UV n, int base, int length);
 extern int  to_digit_string(char *s, UV n, int base, int length);
